+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Security in case of vulnerabilities.
 - Added MAT Transfer creation end point for prepare to complete 
 - Added `Direction To Transfer` task for conversion projects.
 
+### Fixed
+- Create project endpoints no longer throw exception when project group already exists for same UKPRN
+
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-11-06.1066...main) for everything awaiting release
 
 ---